Revert ":bug: :hammer: fixed bug in determine matcher type, lexer token renaming"
This reverts commit e7c3e6b9
