+++ karate-core/src/main/antlr4/com/intuit/karate/core/input.txt
@@ -1,1 +1,1 @@
-driver.eval('{ foo: "(bar)" }') == { hello: 'world' }
+$.addOns[?(@.entitlementStateChangeReason=='RESUBSCRIBE')].addOnOfferID contains only toAddOnOfferIDs
